Add HeaderCdpMessageBar to page stories.
Refactor common page stories components.
